VP9: fix integer overflow in decoder.
The difference of two size_t variables.

+++ vp9/decoder/vp9_decodeframe.c
@@ -1525,10 +1525,10 @@ static int tile_worker_hook(void *arg1, void *arg2) {
 }
 
 // sorts in descending order
-static int compare_tile_buffers(const void *a, const void *b) {
+static int64_t compare_tile_buffers(const void *a, const void *b) {
   const TileBuffer *const buf1 = (const TileBuffer *)a;
   const TileBuffer *const buf2 = (const TileBuffer *)b;
-  return (int)(buf2->size - buf1->size);
+  return (int64_t)buf2->size - (int64_t)buf1->size;
 }
 
 static const uint8_t *decode_tiles_mt(VP9Decoder *pbi, const uint8_t *data,
